@@ -13,11 +13,11 @@
     "@testing-library/user-event": "^12.6.0",
     "cross-env": "^7.0.3",
     "firebase": "^8.2.2",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-images-uploading": "^3.1.2",
     "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^5.3.3",
     "react-scripts": "4.0.1",
     "web-vitals": "^0.2.4"
   },
