@@ -1,24 +1,21 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from 'react-dom/client';
 import { BrowserRouter as Router } from "react-router-dom";
 import App from "./App";
-import reportWebVitals from "./reportWebVitals";
 
 import { StateProvider } from "providers/StateProvider";
 import reducer from "reducers/state_reducer";
 
-ReactDOM.render(
+
+const container = document.getElementById('root');
+const root = createRoot(container);
+
+root.render(
   <React.StrictMode>
     <Router>
       <StateProvider reducer={reducer}>
         <App />
       </StateProvider>
     </Router>
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 );
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
