@@ -1,4 +1,4 @@
-import React, { useEffect, useState, useRef } from "react";
+import React, { useEffect } from "react";
 import "./styles/feed.css";
 // COMPONENTS
 import StoryReel from "../../components/StoryReel/";
@@ -10,70 +10,68 @@ import { useStateValue } from "../../providers/StateProvider";
 import { useApiUtil } from "../../providers/ApiUtil";
 
 function PostFeed({ page, user }) {
-  const [initialRender, setInitialRender] = useState(true);
   const {
     state: {
-      feed: { users, posts, comments },
+      feed: { users, posts, comments, wallId },
     },
   } = useStateValue();
   const { getFeedData } = useApiUtil();
-  const userRef = useRef(user);
+  const changingWall = wallId !== (page === "userWall" ? user.id : page);
 
   useEffect(() => {
-    const numberOfFriends = 10;
-    const getUserIds = (n) => {
-      const shuffled = user.friends.sort(() => 0.5 - Math.random());
-      return user.friends.length > n ? shuffled.slice(0, n) : shuffled;
-    };
+    const setFeedData = (initialRender, numberOfFriends = 10) => {
+      const getUserIds = (n) => {
+        const shuffled = user.friends.sort(() => 0.5 - Math.random());
+        return user.friends.length > n ? shuffled.slice(0, n) : shuffled;
+      };
 
-    const onScroll = async () => {
-      // window.document.body.offsetHeight * .75 <= window.pageYOffset + window.innerHeight
-      if (
-        window.document.body.offsetHeight ===
-          window.pageYOffset + window.innerHeight &&
-        !initialRender
-      ) {
-        getFeedData(
-          getUserIds(numberOfFriends),
-          numberOfFriends,
-          page === "userWall" ? ["wallId", "==", user.id] : null
-        );
-      }
-    };
-
-    if (initialRender || userRef.current !== user) {
       getFeedData(
+        posts,
         getUserIds(numberOfFriends),
         numberOfFriends,
         page === "userWall" ? ["wallId", "==", user.id] : null,
-        true
+        initialRender,
+        page === "userWall" ? user.id : page
       );
-      setInitialRender(false);
-      userRef.current = user;
-    }
+    };
+
+    const onScroll = () => {
+      // window.document.body.offsetHeight * .75 <= window.pageYOffset + window.innerHeight
+      if (
+        window.document.body.offsetHeight ===
+        window.pageYOffset + window.innerHeight
+      ) {
+        if (!changingWall) setFeedData(false);
+      }
+    };
+
+    if (changingWall) setFeedData(true);
+
     window.addEventListener("scroll", onScroll);
     return () => window.removeEventListener("scroll", onScroll);
-  }, [initialRender, user.friends, posts, page, user.id, getFeedData, user]);
+  }, [user.friends, page, user.id, getFeedData, user, posts, changingWall]);
 
-  return initialRender || userRef.current !== user ? (
-    <></>
-  ) : (
+  return (
     <div className="feed">
       <div className="feed__container">
         {page === "userFeed" && <StoryReel />}
         <MessageSender wallId={user.id} />
         {page === "userFeed" && <VideoFeed />}
-        {posts.map((post, idx) => {
-          return (
-            <Post
-              idx={idx}
-              users={users}
-              key={post.id}
-              post={post}
-              comments={comments[post.id]}
-            />
-          );
-        })}
+        {changingWall ? (
+          <></>
+        ) : (
+          posts.map((post, idx) => {
+            return (
+              <Post
+                idx={idx}
+                users={users}
+                key={post.id}
+                post={post}
+                comments={comments[post.id]}
+              />
+            );
+          })
+        )}
         {page === "userFeed" && (
           <div className="feed__noMorePosts">
             <h3>No More Posts</h3>
