@@ -4,6 +4,9 @@ export const actionTypes = {
   SET_LOG_OUT: "SET_LOG_OUT",
   SET_USERS: "SET_USERS",
   SET_POSTS: "SET_POSTS",
+  SET_WALL_ID: "SET_WALL_ID",
+  SET_PROFILE: "SET_PROFILE",
+  SET_PRELOADED_PROFILE: "SET_PRELOADED_PROFILE",
   UPDATE_POST: "UPDATE_POST",
   UPDATE_COMMENTS: "UPDATE_COMMENTS",
   SET_COMMENTS: "SET_COMMENTS",
@@ -93,12 +96,30 @@ const reducer = (state, action) => {
           ...state.feed,
           comments: {
             ...state.feed.comments,
-            [action.postId]: [
-              ...action.comments,
-              ...state.feed.comments[action.postId],
-            ],
-          },
-        },
+            [action.postId]: [...action.comments, ...state.feed.comments[action.postId]]
+          }
+        }
+      };
+
+    case actionTypes.SET_PROFILE:
+        return {
+          ...state,
+          profile: action.profile
+        };
+
+    case actionTypes.SET_PRELOADED_PROFILE:
+      return {
+        ...state,
+        preloadedProfile: action.preloadedProfile
+      };
+
+    case actionTypes.SET_WALL_ID:
+      return {
+        ...state,
+        feed: {
+          ...state.feed,
+         wallId: action.wallId
+        }
       };
 
     case actionTypes.UPDATE_POST:
@@ -106,22 +127,15 @@ const reducer = (state, action) => {
         ...state,
         feed: {
           ...state.feed,
-          ...(action.collection === "posts"
-            ? {
-                posts: state.feed.posts.map((post, idx) =>
-                  idx === action.idx ? action.post : post
-                ),
-              }
+          ...(action.collection === 'posts'
+            ? { posts: state.feed.posts.map((post, idx) => idx === action.idx ? action.post : post )}
             : {
-                comments: {
-                  ...state.feed.comments,
-                  [action.post.postId]: state.feed.comments[
-                    action.post.postId
-                  ].map((comment, idx) =>
-                    idx === action.idx ? action.post : comment
-                  ),
-                },
-              }),
+              comments: {
+                ...state.feed.comments,
+                [action.post.postId]: state.feed.comments[action.post.postId].map((comment, idx) => idx === action.idx ? action.post : comment )
+              }
+            }
+          ),
         },
       };
 
