@@ -1,15 +1,20 @@
-import React from 'react'
-import SidebarRow from './sidebar_row'
-import './styles/sidebar.css'
+import React from "react";
+import SidebarRow from "./sidebar_row";
+import "./styles/sidebar.css";
 
-import { useStateValue } from '../../providers/StateProvider'
+import { useStateValue } from "../../providers/StateProvider";
 
 function Sidebar() {
-  const { state: {user} } = useStateValue()
-  
+  const {
+    state: { user },
+  } = useStateValue();
+
   return (
     <div className="sidebar">
-      <SidebarRow pictureId={user.profilePic} title={`${user.firstName} ${user.lastName}`} />
+      <SidebarRow
+        profilePicData={user.profilePicData}
+        title={`${user.firstName} ${user.lastName}`}
+      />
       <SidebarRow Icon="findFriends" title="Find Friends" />
       <SidebarRow Icon="mostRecent" title="Most Recent" />
       <SidebarRow Icon="favorites" title="Favorites" />
@@ -21,7 +26,7 @@ function Sidebar() {
       <SidebarRow Icon="saved" title="Saved" />
       <SidebarRow Icon="findFriends" title="See More" />
     </div>
-  )
+  );
 }
 
-export default Sidebar
+export default Sidebar;
